Refactor
Removed storage of the menu in a private instance variable.  It's no
longer being used to enable/disable the scan button since the UI is
blocked by the scanning dialog anyway.

Require the location to be accurate within 50 meters before scanning
stops.  This will prevent wild inaccuracies the first time a scan is
performed from the app.